+++ src/MyPluginInfo.cs
@@ -5,6 +5,6 @@
         public const string PLUGIN_GUID = "Scopophobia";
 
         public const string PLUGIN_NAME = "Scopophobia";
-        public const string PLUGIN_VERSION = "1.3.1";
+        public const string PLUGIN_VERSION = "1.3.2";
     }
 }